+++ packages/rpmtw_api_client/lib/src/models/translate/mod_source_info.dart
@@ -3,6 +3,7 @@ import "dart:convert";
 import "package:collection/collection.dart";
 import "package:rpmtw_api_client/src/api_client.dart";
 import "package:rpmtw_api_client/src/models/api_model.dart";
+import 'package:rpmtw_api_client/src/models/list_model_response.dart';
 import "package:rpmtw_api_client/src/models/minecraft/minecraft_mod.dart";
 import "package:rpmtw_api_client/src/models/translate/source_file.dart";
 import "package:rpmtw_api_client/src/models/translate/source_text.dart";
@@ -36,8 +37,9 @@ class ModSourceInfo implements APIModel {
   }
 
   /// [SourceFile] files included in this mod.
-  Future<List<SourceFile>> get files {
-    return SourceFile.list(modSourceInfo: this);
+  Future<ListModelResponse<SourceFile>> getFiles(
+      {int limit = 50, int skip = 0}) {
+    return SourceFile.list(modSourceInfo: this, limit: limit, skip: skip);
   }
 
   Future<List<SourceText>?> get patchouliAddonTexts async {
@@ -118,8 +120,9 @@ class ModSourceInfo implements APIModel {
       RPMTWApiClient.instance.translateResource.getModSourceInfo(uuid);
 
   static Future<ModSourceInfo?> getByMod(MinecraftMod mod) async {
-    List<ModSourceInfo> infos = await RPMTWApiClient.instance.translateResource
-        .listModSourceInfo(mod: mod, limit: 1);
+    List<ModSourceInfo> infos = (await RPMTWApiClient.instance.translateResource
+            .listModSourceInfo(mod: mod, limit: 1))
+        .data;
 
     if (infos.isEmpty) {
       return null;
@@ -129,8 +132,9 @@ class ModSourceInfo implements APIModel {
   }
 
   static Future<ModSourceInfo?> getByNamespace(String namespace) async {
-    List<ModSourceInfo> infos = await RPMTWApiClient.instance.translateResource
-        .listModSourceInfo(namespace: namespace, limit: 1);
+    List<ModSourceInfo> infos = (await RPMTWApiClient.instance.translateResource
+            .listModSourceInfo(namespace: namespace, limit: 1))
+        .data;
 
     if (infos.isEmpty) {
       return null;