+++ packages/rpmtw_api_client/lib/src/models/translate/translation_vote.dart
@@ -2,6 +2,7 @@ import "dart:convert";
 
 import "package:rpmtw_api_client/src/api_client.dart";
 import "package:rpmtw_api_client/src/models/api_model.dart";
+import 'package:rpmtw_api_client/src/models/list_model_response.dart';
 import "package:rpmtw_api_client/src/models/translate/translation.dart";
 
 class TranslationVote implements APIModel {
@@ -86,7 +87,7 @@ class TranslationVote implements APIModel {
   static Future<TranslationVote> getByUUID(String uuid) async =>
       RPMTWApiClient.instance.translateResource.getVote(uuid);
 
-  static Future<List<TranslationVote>> getAllByTranslation(
+  static Future<ListModelResponse<TranslationVote>> getAllByTranslation(
           Translation translation,
           {int limit = 50,
           int skip = 0}) async =>