+++ packages/rpmtw_api_client/lib/src/resources/minecraft_resource.dart
@@ -1,12 +1,13 @@
 import "package:rpmtw_api_client/src/http/api_http_client.dart";
 import "package:rpmtw_api_client/src/http/api_http_response.dart";
+import 'package:rpmtw_api_client/src/models/list_model_response.dart';
 import "package:rpmtw_api_client/src/models/minecraft/minecraft_mod.dart";
 import "package:rpmtw_api_client/src/models/minecraft/minecraft_version_manifest.dart";
 import "package:rpmtw_api_client/src/models/minecraft/mod_integration.dart";
 import "package:rpmtw_api_client/src/models/minecraft/mod_side.dart";
 import "package:rpmtw_api_client/src/models/minecraft/relation_mod.dart";
 import "package:rpmtw_api_client/src/models/minecraft/rpmwiki/wiki_change_log.dart";
-import "package:rpmtw_api_client/src/resources/base_resource.dart";
+import 'package:rpmtw_api_client/src/resources/api_resource.dart';
 import "package:rpmtw_api_client/src/utilities/exceptions.dart";
 
 class MinecraftResource extends APIResource {
@@ -172,7 +173,7 @@ class MinecraftResource extends APIResource {
   /// 透過 模組名稱/模組譯名/模組 ID 來搜尋 Minecraft 模組
   /// [limit] 取得的資料數量 (預設為 50，最大為 50)
   /// [skip] 跳過的資料數量 (預設為 0)
-  Future<List<MinecraftMod>> search(
+  Future<ListModelResponse<MinecraftMod>> search(
       {String? filter,
       int? limit,
       int? skip,
@@ -187,9 +188,7 @@ class MinecraftResource extends APIResource {
 
     Map data = response.data;
 
-    return List<MinecraftMod>.from((data["mods"] as List)
-        .cast<Map<String, dynamic>>()
-        .map((e) => MinecraftMod.fromMap(e)));
+    return ListModelResponse.fromMap<MinecraftMod>(data);
   }
 
   /// 查詢 RPMWiki 的變更日誌
@@ -197,7 +196,7 @@ class MinecraftResource extends APIResource {
   /// [skip] 跳過的資料數量 (預設為 0)
   /// [dataUUID] 要篩選的資料 UUID
   /// [userUUID] 要篩選的資料編輯者 UUID
-  Future<List<WikiChangelog>> filterChangelogs(
+  Future<ListModelResponse<WikiChangelog>> filterChangelogs(
       {int? limit, int? skip, String? dataUUID, String? userUUID}) async {
     APIHttpResponse response =
         await httpClient.get("/minecraft/changelog", query: {
@@ -208,9 +207,8 @@ class MinecraftResource extends APIResource {
     });
 
     Map data = response.data;
-    return List<WikiChangelog>.from((data["changelogs"] as List)
-        .cast<Map<String, dynamic>>()
-        .map((e) => WikiChangelog.fromMap(e)));
+
+    return ListModelResponse.fromMap<WikiChangelog>(data);
   }
 
   /// 取得 Minecraft 版本資訊 (由 Mojang API 提供，RPMTW API 伺服器每天快取一次資料)