+++ packages/rpmtw_api_client/lib/src/resources/other_resource.dart
@@ -1,6 +1,6 @@
 import "package:rpmtw_api_client/src/http/api_http_client.dart";
 import "package:rpmtw_api_client/src/http/api_http_response.dart";
-import "package:rpmtw_api_client/src/resources/base_resource.dart";
+import 'package:rpmtw_api_client/src/resources/api_resource.dart';
 
 class OtherResource extends APIResource {
   const OtherResource(APIHttpClient httpClient) : super(httpClient);